refactor: componentized main email template management component for organization. ngl kinda a lot of components now but a lot more clean
