+++ client/src/components/emailTemplateManagement/components/index.js
@@ -1,0 +1,10 @@
+export { SearchBar } from "./SearchBar";
+export { BreadcrumbNav } from "./BreadcrumbNav";
+export { Pagination } from "./Pagination";
+export { FolderCard } from "./FolderCard";
+export { TemplateCard } from "./TemplateCard";
+export { NewTemplatePopover } from "./NewTemplatePopover";
+export { NewFolderPopover } from "./NewFolderPopover";
+export { EmptyFolderState } from "./EmptyFolderState";
+export { Sidebar } from "./Sidebar";
+export { NewTemplateSection } from "./NewTemplateSection";
